api/requirements.txt
flask
flask-cors
mysql-connector-python
requests
spotipy
dotenv
flask-session
supabase
pyjwt
datetime
openai
waitress